+++ Packages/com.mimylab.fukuroudon/Runtime/PlayerAudioMaster/Scripts/PlayerAudioRegulatorBase.cs
@@ -1,0 +1,24 @@
+﻿/*
+Copyright (c) 2023 Mimy Quality
+Released under the MIT license
+https://opensource.org/licenses/mit-license.php
+*/
+
+namespace MimyLab
+{
+    using UdonSharp;
+    using UnityEngine;
+    using VRC.SDKBase;
+    //using VRC.Udon;
+    //using VRC.SDK3.Components;
+
+    [AddComponentMenu("Fukuro Udon/PlayerAudio Master/PA Regulator Base")]
+    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
+    public class PlayerAudioRegulatorBase : IPlayerAudioRegulator
+    {
+        protected override bool CheckApplicableInternal(VRCPlayerApi target)
+        {
+            return true;
+        }
+    }
+}